move tw-statistic.js into statistic.js
