+++ src/components/blocks/home/homeStats.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { useStaticQuery, graphql } from "gatsby"
-import Statistic from "../../shared/tw-statistic"
+import Statistic from "../../shared/statistic"
 import { Container } from "react-bootstrap"
 
 const query = graphql`
@@ -19,11 +19,11 @@ export default function HomeStats() {
   const stats = useStaticQuery(query).blockYaml.stats
   return (
     <>
-    <Container className="content-block">
-      <h2 className="mt-5 mb-5">How We Rank Among the World</h2>
+    <div className="tw-container tw-content-block">
+      <h2 className="tw-my-12">How We Rank Among the World</h2>
 
-      <Statistic.Gradient stats={stats} />
-    </Container>
+      <Statistic.Gradient stats={stats} fullWidthBG={true} />
+    </div>
     </>
   )
 }
